+++ now/dialog.py
@@ -15,7 +15,7 @@ from now.common import options
 from now.common.options import construct_app
 from now.constants import MODALITY_TO_MODELS, Apps, DialogStatus
 from now.now_dataclasses import DialogOptions, UserInput
-from now.run_all_k8s import maybe_prompt_user
+from now.thirdparty.PyInquirer.prompt import prompt
 from now.utils import DemoAvailableException, RetryException
 
 cur_dir = pathlib.Path(__file__).parent.resolve()
@@ -94,7 +94,7 @@ def configure_option(
         option.choices = option.choices(user_input, **kwargs)
 
     while True:
-        val = now.utils.prompt_value(
+        val = prompt_value(
             **option.__dict__,
             **kwargs,
         )
@@ -130,3 +130,21 @@ def prompt_value(
     if choices is not None:
         qs['choices'] = choices
     return maybe_prompt_user(qs, name, **kwargs)
+
+
+def maybe_prompt_user(questions, attribute, **kwargs):
+    """
+    Checks the `kwargs` for the `attribute` name. If present, the value is returned directly.
+    If not, the user is prompted via the cmd-line using the `questions` argument.
+
+    :param questions: A dictionary that is passed to `PyInquirer.prompt`
+        See docs: https://github.com/CITGuru/PyInquirer#documentation
+    :param attribute: Name of the value to get. Make sure this matches the name in `kwargs`
+
+    :return: A single value of either from `kwargs` or the user cli input.
+    """
+    if kwargs and attribute in kwargs:
+        return kwargs[attribute]
+    else:
+        answer = prompt(questions)
+        return answer[attribute]
